[#64570338] Fix session id bug.
Session id bug seems to appear in a multithreading environment when
another thread reuse an open connection and reset the current_session_id
useb by a previous thread.

This commit fix the bug by setting again the current_session_id at the
beginning of a transaction. Hopefully, a SQL connection shouldn't be
reused by Sequel in the middle of a transaction.